Move the immediate visitor notification out of publish_visit
publish_visit is about pushing live updates over the chat bus. Hanging a
notification off it meant that function now decided who to notify, fetched
a row it didn't need for its own work, and grew a flag to keep the two
apart -- and its shape dictated when the notification could happen at all.

The notification lives in its own module and is called alongside
publish_visit rather than from inside it, so neither knows about the
other. It asks the cheap question first (does this person want to hear
about visitors immediately) and only then reads the prospect's own
visitors-tab item, which is still the single source of truth for whether
a visit is visible and what the visitor is called.

The cost of the split is that a prospect who is both online and on
"Immediately" has that item read twice, once for each purpose. That pair
is rare, and the alternative was one function doing both jobs.

visitorspush is now byte-identical to main again.
